HERA-scraper/src/test/java/be/rommens/hera/providers/ReadComicsScraperTest.java
package be.rommens.hera.providers;

/*import com.github.tomakehurst.wiremock.WireMockServer;
import org.hamcrest.Matchers;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;*/
import com.github.tomakehurst.wiremock.WireMockServer;
import org.hamcrest.Matchers;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

//import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
import java.io.IOException;

import static com.github.tomakehurst.wiremock.client.WireMock.*;
import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

/**
 * User : cederik
 * Date : 29/03/2020
 * Time : 14:22
 */
@SpringBootTest(classes = ReadComicsScraper.class)
public class ReadComicsScraperTest {

    public static WireMockServer wireMockServer = new WireMockServer(options().port(8888));

    @BeforeAll
    static void setupClass() {
        wireMockServer.start();
    }

    @AfterEach
    void after() {
        wireMockServer.resetAll();
    }

    @AfterAll
    static void clean() {
        wireMockServer.shutdown();
    }

    @Autowired
    private ReadComicsScraper readComicsScraper;

    @Test
    public void test() {
        assertThat(readComicsScraper.getProviderProperty(), is("http://readcomicsonline.ru/comic/"));
    }

    @Test
    public void testW() throws IOException {
        String result = readComicsScraper.testConn();
        assertThat(result, Matchers.containsString("world"));
    }
}
